More rest fixes
