+++ package.json
@@ -10,21 +10,24 @@
   "scripts": {
     "server": "nodemon server/index.js",
     "build": "webpack --mode production --watch",
-    "test": "jest tests/test.js"
+    "test": "jest",
+    "testclient": "__test__/App.test.js",
+    "testserver": " __test__/index.test.js"
   },
   "dependencies": {
     "axios": "0.19.2",
     "bluebird": "^3.7.2",
     "body-parser": "^1.19.0",
     "dotenv": "8.2.0",
+    "enzyme-adapter-react-15.4": "^1.4.1",
     "express": "^4.15.0",
     "jquery": "^3.1.1",
     "mysql": "2.18.1",
     "mysql2": "2.1.0",
     "nodemon": "2.0.2",
     "react": "^15.4.2",
+    "react-addons-test-utils": "^15.0.0",
     "react-dom": "^15.4.2",
-    "react-test-renderer": "^16.13.1",
     "request": "^2.81.0",
     "sequelize": "5.21.5"
   },
@@ -39,6 +42,8 @@
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "faker": "4.1.0",
     "jest": "^25.1.0",
+    "react-dom": "^15.4.2",
+    "react-test-renderer": "^15.4.1",
     "supertest": "^4.0.2",
     "webpack": "4.42.0",
     "webpack-cli": "3.3.11"
